requirements.txt
transformers>=4.30.0
trl~=0.15.1
datasets>=2.10.0
peft~=0.14.0
accelerate


torch~=2.5.1+cu121
unsloth~=2025.2.15
numpy~=1.26.4
packaging~=24.2
vllm~=0.7.2.dev56+gbf3b79ef
huggingface-hub~=0.28.1
pandas~=2.2.3
python-dotenv~=1.0.1
plotly~=6.0.0
tqdm~=4.67.1